fix: use "size_t"

+++ include/pt/algorithm/blas.h
@@ -6,7 +6,7 @@
 
 extern void dscal(int* n, double* a, double* x, int* inc_x);
 
-pt_error pt_dscal(size_t n, double a, double* x, int inc_x);
+pt_error pt_dscal(size_t n, double a, double* x, size_t inc_x);
 
 #define pt_scal(_n, _a, _x, _inc_x) \
   _Generic((_a), double: pt_dscal)((_n), (_a), (_x), (_inc_x))
